Improved collision, detection, added new levels, removed animation component, and a few other changes.

+++ Src/ECS/ColliderComponent.h
@@ -8,7 +8,6 @@ class ColliderComponent : public Component {
 public:
 	SDL_Rect collider;
 	std::string tag;
-
 	TransformComponent* transform;
 
 	ColliderComponent(std::string t) {
@@ -28,7 +27,7 @@ public:
 	void update() override {
 		collider.x = static_cast<int>(transform->position.x);
 		collider.y = static_cast<int>(transform->position.y);
-		collider.w = transform->width * transform->scale;
-		collider.h = transform->height * transform->scale;
+		collider.w = static_cast<int>(transform->width * transform->scale);
+		collider.h = static_cast<int>(transform->height * transform->scale);
 	}
 };
